# api/docs/openapi.yml
openapi: "3.1.0"
info:
  title: Taillight API
  description: |
    Taillight is a centralized srvlog and application-log viewer with real-time
    streaming, analysis reports, and rsyslog/Juniper integration.
  version: "1.0.0"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: Default API prefix
...
security:
  - session: []
  - apiKey: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /notifications/channels:
    get:
      operationId: listChannels
      tags: [Notifications]
      summary: List notification channels
      responses:
        "200":
          description: Channel list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/NotificationChannel"
    post:
      operationId: createChannel
      tags: [Notifications]
      summary: Create notification channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationChannelInput"
      responses:
        "201":
          description: Channel created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/NotificationChannel"
        "400":
          $ref: "#/components/responses/BadRequest"
  /notifications/channels/{id}:
    get:
      operationId: getChannel
      tags: [Notifications]
      summary: Get notification channel
      parameters:
        - $ref: "#/components/parameters/IntID"
      responses:
        "200":
          description: Channel detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/NotificationChannel"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateChannel
      tags: [Notifications]
      summary: Update notification channel
      parameters:
        - $ref: "#/components/parameters/IntID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationChannelInput"
      responses:
        "200":
          description: Channel updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/NotificationChannel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteChannel
      tags: [Notifications]
      summary: Delete notification channel
      parameters:
        - $ref: "#/components/parameters/IntID"
      responses:
        "204":
          description: Channel deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /notifications/channels/{id}/test:
    post:
      operationId: testChannel
      tags: [Notifications]
      summary: Send test notification
      parameters:
        - $ref: "#/components/parameters/IntID"
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestChannelResult"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          description: Notification engine not enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /notifications/rules:
    get:
      operationId: listRules
      tags: [Notifications]
      summary: List notification rules
      responses:
        "200":
          description: Rule list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/NotificationRule"
    post:
      operationId: createRule
      tags: [Notifications]
      summary: Create notification rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationRuleInput"
      responses:
        "201":
          description: Rule created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/NotificationRule"
        "400":
          $ref: "#/components/responses/BadRequest"
  /notifications/rules/{id}:
    get:
      operationId: getRule
      tags: [Notifications]
      summary: Get notification rule
      parameters:
        - $ref: "#/components/parameters/IntID"
      responses:
        "200":
          description: Rule detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/NotificationRule"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateRule
      tags: [Notifications]
      summary: Update notification rule
      parameters:
        - $ref: "#/components/parameters/IntID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationRuleInput"
      responses:
        "200":
          description: Rule updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/NotificationRule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteRule
      tags: [Notifications]
      summary: Delete notification rule
      parameters:
        - $ref: "#/components/parameters/IntID"
      responses:
        "204":
          description: Rule deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /notifications/log:
    get:
      operationId: listNotificationLog
      tags: [Notifications]
      summary: List notification delivery log
      parameters:
        - name: rule_id
          in: query
          schema:
            type: integer
          description: Filter by rule ID
        - name: channel_id
          in: query
          schema:
            type: integer
          description: Filter by channel ID
        - name: status
          in: query
          schema:
            type: string
          description: Filter by delivery status
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: Log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/NotificationLogEntry"

  # ── Device ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: tl_session
      description: Session cookie set by POST /auth/login
    apiKey:
      type: http
      scheme: bearer
      description: "API key obtained from POST /auth/keys"

  # ── Parameters ──────────────────────────────────────────────────────────
  parameters:
...
    IntID:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    From:
      name: from
      in: query
      schema:
        type: string
        format: date-time
      description: Start time (inclusive)
    To:
      name: to
      in: query
      schema:
        type: string
        format: date-time
      description: End time (inclusive)
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
    NotificationChannel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [slack, webhook, email]
        config:
          type: object
          additionalProperties: true
          description: Backend-specific configuration (webhook URL, Slack token, SMTP settings, etc.)
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    NotificationChannelInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [slack, webhook, email]
        config:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
    NotificationRule:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        enabled:
          type: boolean
        event_kind:
          type: string
          enum: [srvlog, applog]
        hostname:
          type: string
        programname:
          type: string
        severity:
          type: integer
          nullable: true
        severity_max:
          type: integer
          nullable: true
        facility:
          type: integer
          nullable: true
        syslogtag:
          type: string
        msgid:
          type: string
        service:
          type: string
        component:
          type: string
        host:
          type: string
        level:
          type: string
        search:
          type: string
        channel_ids:
          type: array
          items:
            type: integer
        burst_window:
          type: integer
          description: Number of events before triggering a burst notification
        cooldown_seconds:
          type: integer
          description: Minimum seconds between notifications for this rule
        group_by:
          type: string
          description: Field to group events by (e.g. hostname, service)
        max_cooldown_seconds:
          type: integer
          description: Maximum cooldown after exponential backoff
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    NotificationRuleInput:
      type: object
      required: [name, event_kind]
      properties:
        name:
          type: string
        enabled:
          type: boolean
        event_kind:
          type: string
          enum: [srvlog, applog]
        hostname:
          type: string
        programname:
          type: string
        severity:
          type: integer
          nullable: true
        severity_max:
          type: integer
          nullable: true
        facility:
          type: integer
          nullable: true
        syslogtag:
          type: string
        msgid:
          type: string
        service:
          type: string
        component:
          type: string
        host:
          type: string
        level:
          type: string
        search:
          type: string
        channel_ids:
          type: array
          items:
            type: integer
        burst_window:
          type: integer
        cooldown_seconds:
          type: integer
        group_by:
          type: string
        max_cooldown_seconds:
          type: integer
    NotificationLogEntry:
      type: object
      properties:
        id:
          type: integer
        created_at:
          type: string
          format: date-time
        rule_id:
          type: integer
        channel_id:
          type: integer
        event_kind:
          type: string
        event_id:
          type: integer
        status:
          type: string
        reason:
          type: string
          nullable: true
        event_count:
          type: integer
        status_code:
          type: integer
          nullable: true
        duration_ms:
          type: integer
        payload:
          type: object
          nullable: true
          additionalProperties: true
    TestChannelResult:
      type: object
      properties:
        success:
          type: boolean
        status_code:
          type: integer
        error:
          type: string
          nullable: true
        duration_ms:
          type: integer

    # ── Device ───────────────────────────────────────────────────────────
...